+++ core/client/index.js
@@ -5,8 +5,8 @@
  ************
  ************
  */
-import { OlmWrapper } from "./crypto/olmWrapper.js";
-import { ServerComm } from "./serverComm/socketIOWrapper.js";
+import { OlmWrapper } from "./olmWrapper.js";
+import { ServerComm } from "./serverComm.js";
 export class Core {
     olmWrapper;
     #serverComm;
