+++ higher/db/localStorageWrapper.js
