+++ higher/index.ts
@@ -23,8 +23,8 @@
 // TODO need new special name for LINKED group (confusing when linking non-LINKED groups)
 
 import EventEmitter from "events";
-import { Core } from "../core/client/core.js";
-import { LocalStorageWrapper } from "../core/client/db/localStorageWrapper.js";
+import { Core } from "../core/client";
+import { LocalStorageWrapper } from "./db/localStorageWrapper.js";
 
 /* doubly-linked tree, allows cycles */
 const NAME         : string = "name";
